sched: fix timer state machine
Timers currently have two states: expired, and not expired.  But how is
cancel() to distinguish between an armed-but-not-expired timer, and a timer
which was never armed (using set())?  It can't, and will probably crash.

Fix by adding a state to the timer.  Now there are three distinct states:
free, armed, and expired, corresponding to the timer life cycle.