sched: simplify vruntime for currently running thread
Since vruntime needs to be advanced by (scaled) running time, when a thread
starts running, we subtract the start clock value (t1), and when it stops,
we add the stop clock value (t2), with the end result that we add t2-t1, or
the running time.

While this works, it makes the code less readable, since the meaning of
_vruntime is drastically different for running threads - its the vruntime
minus the reading of the clock at some point in the past.

Simplify by not doing this, and instead keeping the start time in a separate
field (running_since).

Note _vruntime for a running thread is still subtly different from that of a
stopped thread, since we need to add this delta time.