sched: adjust timer expiration when the clock is not running
Change the condition for expiration to allow an exact match between the
current time and the timer expiration time.

This helps the scheduler keep going when the clock is still not running
during system startup.  Ideally we shouldn't depend on such details, but
fixing this is too complicated for now.