cli.js: cast command line arguments to JS strings
mainargs is a String[] set by Java (see RhinoCLI.java), some of the commands
that access the elements of this array fail because the strings are Java strings
and not Javascript strings.

this patch perform a proper cast on init(), just before invoking the command.

now it's also possible to start a test from the command line, like so:

$ sudo ./scripts/run.py -n -e "java.so -jar /java/cli.jar test <test-name>" -c2 -m1G