provide C symbols for {f}statvfs64
We need to provide C linkage for those as well. It should really be in the header
file, but the fact that we are defining this as a weak symbol is confusing the compiler.

While we are on it, fstatvfs64 is aliasing to statvfs instead of fstatvfs. It doesn't seem
to be what we want.