xen: pad xen_shared_info to a page in size
Xen will map a page here, but the structure is not a full page in size. This is
true for both our structure, and the imported structure, which is about 3k in size.

This is very unfortunate, since data that is put right after it will end up being
overwritten. Pad it properly to make sure it does not happen.